pyproject.toml
[tool.poetry]
name = "cafe-orders"
version = "1.0.0"
description = "Система управления заказами в кафе"
authors = ["i.sabirov <i.sabirov@medrc.ru>"]
readme = "README.md"
license = "ISC"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.14,<4.0"
fastapi = ">=0.115.0"
uvicorn = {extras = ["standard"], version = ">=0.30.0"}
sqlalchemy = ">=2.0.30"
alembic = ">=1.13.0"
pydantic = ">=2.9.0"
pydantic-settings = ">=2.6.0"
python-jose = {extras = ["cryptography"], version = ">=3.3.0"}
bcrypt = ">=4.0.0"
python-multipart = ">=0.0.9"
aiosqlite = ">=0.20.0"
email-validator = ">=2.1.0"
psycopg2-binary = ">=2.9.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.2.0"
pytest-asyncio = ">=0.23.6"
pytest-cov = ">=5.0.0"
httpx = ">=0.27.0"
playwright = ">=1.40.0"
psycopg2-binary = ">=2.9.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
